refactor(desktop): unify OAuth dialogs into shared OAuthDialog
Extract shared OAuthDialog component with provider config object.
AnthropicOAuthDialog and OpenAIOAuthDialog become thin wrappers
that pass provider-specific labels and options.
