+++ apps/desktop/src/renderer/components/Chat/ChatInterface/components/ModelPicker/components/OAuthDialog/index.ts
@@ -1,0 +1,1 @@
+export { OAuthDialog, type OAuthDialogProps } from "./OAuthDialog";
